Refactor and fix data
Added a scripts/lib folder to store common/utility scripts that
are included in other scripts (but not executed by inspector).
Refactored so that the thinpool stats are saved in their own .stdout
file to be looked at later.
